Optimize Storage of Palette Index
Changed paletteIndex and palette variables from int to uint8_t, as the number of palettes will remain well below 255. This should reduce our memory usage by 6 bytes.

+++ src/palettes.c
@@ -364,7 +364,7 @@
 };
 
 // Implementation of the Palette function
-void Palette(int paletteIndex, int iterations, int *y, int *u, int *v)
+void Palette(uint8_t paletteIndex, int iterations, int *y, int *u, int *v)
 {
   uint8_t (*p)[256][3];  // Pointer to the selected palette
